Ensure the Intel library files are not modified
Each output's build section overrides the overall "build" section,
resetting options to the defaults.

Because of this the Intel daal library is being run through patchelf
to set an RPATH, because that's the default behaviour.

The packaged .so files are then no longer bitwise identical to the
Intel original, a violation of the Intel licence, which requires the
software to be distributed "without modification".